Add support for reading/writing/deleting multiple files at once

+++ index.js
@@ -4136,6 +4136,15 @@ class FileSystem {
       } else {
         this._rm = rmRecursive.bind(null, this);
       }
+      if (fs.promises.readFiles) {
+        this._readFiles = fs.promises.readFiles.bind(fs.promises);
+      }
+      if (fs.promises.writeFiles) {
+        this._writeFiles = fs.promises.writeFiles.bind(fs.promises);
+      }
+      if (fs.promises.unlinkMany) {
+        this._unlinkMany = fs.promises.unlinkMany.bind(fs.promises);
+      }
       this._rmdir = fs.promises.rmdir.bind(fs.promises);
       this._unlink = fs.promises.unlink.bind(fs.promises);
       this._stat = fs.promises.stat.bind(fs.promises);
@@ -4203,6 +4212,21 @@ class FileSystem {
       return null
     }
   }
+    /**
+   * 
+   * @param {array} filepaths 
+   * @param {object|string} options 
+   * @returns {Promise<Buffer[]|string[]|null[]>}
+   */
+    async readFiles(filepaths, options = {}) {
+      const reads = await this._readFiles(filepaths);
+      return reads.map((read) => {
+        if (typeof read !== "string") {
+          return Buffer.from(read)
+        }
+        return read;
+      });
+  }
 
   /**
    * Write a file (creating missing directories if need be) without throwing errors.
@@ -4220,6 +4244,17 @@ class FileSystem {
       await this.mkdir(dirname(filepath));
       await this._writeFile(filepath, contents, options);
     }
+  }
+
+  /**
+   * 
+   * @param {[filepath: string, content: Buffer|Uint8Array|string]} filepathsAndData 
+   * @param {object|string} options 
+   * @returns 
+   */
+  async writeFiles(filepathsAndData, options = {}) {
+      await this._writeFiles(filepathsAndData, options);
+      return;
   }
 
   /**
@@ -4258,6 +4293,17 @@ class FileSystem {
       if (err.code !== 'ENOENT') throw err
     }
   }
+
+    /**
+   * Delete a file without throwing an error if it is already deleted.
+   */
+    async rmMany(filepaths) {
+      try {
+        await this._unlinkMany(filepaths);
+      } catch (err) {
+        if (err.code !== 'ENOENT') throw err
+      }
+    }
 
   /**
    * Delete a directory without throwing an error if it is already deleted.
@@ -5686,6 +5732,9 @@ async function _checkout({
     let ops;
     // First pass - just analyze files (not directories) and figure out what needs to be done
     try {
+      if(onProgress) {
+        await onProgress({ total: 0, phase: "pre-analyze", loaded: 0});
+      }
       ops = await analyze({
         fs,
         cache,
@@ -5704,7 +5753,10 @@ async function _checkout({
         throw err
       }
     }
-
+    if(onProgress) {
+      await onProgress({ total: 0, phase: "post-analyze", loaded: 0});
+    }
+  
     // Report conflicts
     const conflicts = ops
       .filter(([method]) => method === 'conflict')
@@ -5733,21 +5785,27 @@ async function _checkout({
 
     let count = 0;
     const total = ops.length;
+    //if we're going to do a majority of just pure file writes/updates, then lets read
     await GitIndexManager.acquire({ fs, gitdir, cache }, async function(index) {
+      //delete many only when fs has correct extra method
+      if (fs._unlinkMany) {
+        const deleteOps = ops.filter(([method])=> method === "delete").map(([method, fullpath]) => `${dir}/${fullpath}`);
+        await fs.rmMany(deleteOps);
+      }
       await Promise.all(
         ops
           .filter(
             ([method]) => method === 'delete' || method === 'delete-index'
           )
           .map(async function([method, fullpath]) {
-            const filepath = `${dir}/${fullpath}`;
-            if (method === 'delete') {
-              await fs.rm(filepath);
+            if (!fs._unlinkMany && method === 'delete') {	
+              const filepath = `${dir}/${fullpath}`;	
+              await fs.rm(filepath);	
             }
             index.delete({ filepath: fullpath });
             if (onProgress) {
               await onProgress({
-                phase: 'Updating workdir',
+                phase: 'Updating workdir: rm',
                 loaded: ++count,
                 total,
               });
@@ -5768,7 +5826,7 @@ async function _checkout({
             await fs.rmdir(filepath);
             if (onProgress) {
               await onProgress({
-                phase: 'Updating workdir',
+                phase: 'Updating workdir: rmdir',
                 loaded: ++count,
                 total,
               });
@@ -5794,7 +5852,7 @@ async function _checkout({
           await fs.mkdir(filepath);
           if (onProgress) {
             await onProgress({
-              phase: 'Updating workdir',
+              phase: 'Updating workdir: mkdir',
               loaded: ++count,
               total,
             });
@@ -5803,6 +5861,53 @@ async function _checkout({
     );
 
     await GitIndexManager.acquire({ fs, gitdir, cache }, async function(index) {
+      //only execute this enhanced performance methodology if our fs has the required internal functions, otherwise run the standard path
+      if (fs._writeFiles && fs._unlinkMany) {
+        const writeOps = ops.filter(([method]) => method === "create" || method === "update");
+        const deletes = [];
+        const modeWrites = [];
+        const symlinkWrites = [];
+        const regularWrites = [];
+        for (const [_method, fullpath, oid, mode, chmod] of writeOps) {
+          const filepath = `${dir}/${fullpath}`;
+          if (chmod) {
+            deletes.push(filepath);
+          }
+          const { object } = await _readObject({ fs, cache, gitdir, oid });
+          const write = [filepath, object];
+          if (mode === 0o100644) {
+            regularWrites.push(write);
+          } else if (mode === 0o100755) {
+            modeWrites.push(write);
+          } else if (mode === 0o120000) {
+            symlinkWrites.push(write);
+          } else {
+            throw new InternalError(
+              `Invalid mode 0o${mode.toString(8)} detected in blob ${oid}`
+            )
+          }
+        }
+  
+        await fs.rmMany(deletes);
+        if (onProgress) {
+          await onProgress({ loaded: 0, total: 0, phase: "deleted files for chmod reasons"});
+        }
+  
+        await fs.writeFiles(regularWrites, {});
+        if (onProgress) {
+          await onProgress({ loaded: 0, total: regularWrites.length, phase: "wrote regular files"});
+        }
+        await fs.writeFiles(modeWrites, { mode: 0o777 });
+        if (onProgress) {
+          await onProgress({ loaded: 0, total: modeWrites.length, phase: "wrote mode files"});
+        }
+        await Promise.all(symlinkWrites.map(([filepath, data]) => fs.writelink(filepath, data)));
+        if (onProgress) {
+          await onProgress({ loaded: 0, total: symlinkWrites.length, phase: "wrote symlink files"});
+        }
+
+      }
+
       await Promise.all(
         ops
           .filter(
@@ -5815,7 +5920,7 @@ async function _checkout({
           .map(async function([method, fullpath, oid, mode, chmod]) {
             const filepath = `${dir}/${fullpath}`;
             try {
-              if (method !== 'create-index' && method !== 'mkdir-index') {
+              if (!fs._writeFiles && method !== 'create-index' && method !== 'mkdir-index') {
                 const { object } = await _readObject({ fs, cache, gitdir, oid });
                 if (chmod) {
                   // Note: the mode option of fs.write only works when creating files,
@@ -5838,7 +5943,6 @@ async function _checkout({
                   )
                 }
               }
-
               const stats = await fs.lstat(filepath);
               // We can't trust the executable bit returned by lstat on Windows,
               // so we need to preserve this value from the TREE.
@@ -5855,9 +5959,10 @@ async function _checkout({
                 stats,
                 oid,
               });
+
               if (onProgress) {
                 await onProgress({
-                  phase: 'Updating workdir',
+                  phase: 'Updating workdir: write post',
                   loaded: ++count,
                   total,
                 });
@@ -5886,6 +5991,29 @@ async function _checkout({
     }
   }
 }
+
+async function readAllFiles({
+  fs,
+  gitdir,
+  ref,
+}) {
+  const cache = {};
+  const treeResult = await _readTree({ fs, gitdir, oid: ref, cache });
+  return _walk({ fs, gitdir, cache, trees: [TREE({ref: treeResult.oid})], map: async (fileName, entries) => {
+    const fileReadResult = { filePath: fileName, fileData: ""};
+        //if there's content, use it, otherwise return an "empty" FileReadResult
+        if (entries[0]) {
+          const content = await entries[0].content();
+          if (content && typeof content === "object") {
+            fileReadResult.fileData = content;
+          } else {
+            fileReadResult.type = "Directory";
+          }
+        }
+        return fileReadResult;
+  }});
+}
+
 
 async function analyze({
   fs,
